% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_articles_to_df.R
\name{all_articles_to_df}
\alias{all_articles_to_df}
\title{Convert list of PubMed XLM files to dataframe}
\usage{
all_articles_to_df(d.fls)
}
\arguments{
\item{d.fls}{The list of XLM PubMed data.}
}
\description{
Convert list of PubMed XLM files to dataframe
}
