% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_to_df2.R
\name{article_to_df2}
\alias{article_to_df2}
\title{Convert PubMed XLM files to dataframe}
\usage{
article_to_df2(
  pubmedArticle,
  autofill = FALSE,
  max_chars = 500,
  getKeywords = FALSE,
  getAuthors = TRUE
)
}
\arguments{
\item{pubmedArticle}{The PubMed data, in XLM form.}

\item{autofill}{Autofill the affiliation address when missing.}

\item{max_chars}{Maximum number of characters for the address.}

\item{getKeywords}{Whether to extract keywords as well.}

\item{getAuthors}{Whether to extract the author information.}
}
\description{
Convert PubMed XLM files to dataframe
}
\examples{
\dontrun{
dami_query <- paste(
    "passion [Title/Abstract]",
    "AND Dualistic Model of Passion [Text Word]",
    "AND ('2023/01/01' [Date - Publication] : '2023/12/31' [Date - Publication])"
   )
dami_on_pubmed <- easyPubMed::get_pubmed_ids(dami_query)
dami_abstracts_xml <- easyPubMed::fetch_pubmed_data(dami_on_pubmed)
dami_abstracts_list <- easyPubMed::articles_to_list(dami_abstracts_xml)
article_to_df2(pubmedArticle = dami_abstracts_list[[2]], autofill = FALSE)[1,]
}
}
