% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_pubmed_download2.R
\name{batch_pubmed_download2}
\alias{batch_pubmed_download2}
\title{Download PubMed data with query string}
\usage{
batch_pubmed_download2(
  pubmed_query_string,
  year_low = 2023,
  year_high = 2030,
  data_folder = "data",
  batch_size = 5000,
  api_key = NULL
)
}
\arguments{
\item{pubmed_query_string}{The PubMed query string.}

\item{year_low}{The year the data should start.}

\item{year_high}{The year the data should end.}

\item{data_folder}{Where to save the data.}

\item{batch_size}{The download batch size.}

\item{api_key}{The api key for faster processing (optional).}
}
\description{
Download PubMed data with query string
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
\dontrun{
pubmed_query_string <- paste(
  "passion [Title/Abstract]",
  "AND Dualistic Model of Passion [Text Word]"
)

batch_pubmed_download2(
  pubmed_query_string,
  year_low = 2023,
  year_high = 2023
)
}
\dontshow{
setwd(.old_wd)
}
}
