% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pubmed_api_token.R
\name{check_pubmed_api_token}
\alias{check_pubmed_api_token}
\title{Check whether the PubMed API token is of the correct length and valid}
\usage{
check_pubmed_api_token(API_TOKEN_PUBMED)
}
\arguments{
\item{API_TOKEN_PUBMED}{The PubMed API token}
}
\description{
Check whether the PubMed API token is of the correct length and valid
}
\examples{
API_TOKEN_PUBMED <- c(
  GNZuCGNZuCGNZuCGNZuCGNZuCtwed3twed32
)

check_pubmed_api_token(API_TOKEN_PUBMED)
}
