% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_journals_continents.R
\name{clean_journals_continents}
\alias{clean_journals_continents}
\title{Clean dataframe, for names of journals and continents}
\usage{
clean_journals_continents(data)
}
\arguments{
\item{data}{The processed dataframe of data}
}
\description{
Clean dataframe, for names of journals and continents
}
