% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hex_to_char.R
\name{convert_hex_to_char}
\alias{convert_hex_to_char}
\title{Convert a Hex string to regular text}
\usage{
convert_hex_to_char(hex_string)
}
\arguments{
\item{hex_string}{The address containg hex characters to convert.}
}
\description{
Convert a Hex string to regular text
}
\examples{
address <- c(
  "D&#xe9;partement de Psychologie, Universit&#xe9; du Qu&#xe9;bec &#xe0; Montr&#xe9;al",
  "Department of Behavioural and Cognitive Sciences, University of Luxembourg"
)
convert_hex_to_char(address)

}
