% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\docType{data}
\name{countries}
\alias{countries}
\title{List of countries taken from the package \code{countrycode}}
\format{
A vector of class character containing country names.
}
\usage{
countries
}
\description{
List of countries taken from the package \code{countrycode}
}
\keyword{datasets}
