% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_journals_continents.R
\name{detect_missing_journals}
\alias{detect_missing_journals}
\title{Detect missing journals}
\usage{
detect_missing_journals(data)
}
\arguments{
\item{data}{The processed dataframe of data}
}
\description{
Detect missing journals
}
