% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph_country_year.R
\name{dygraph_country_year}
\alias{dygraph_country_year}
\title{Generate a dygraph of journal paper percentages, by country and year}
\usage{
dygraph_country_year(data)
}
\arguments{
\item{data}{The processed dataframe of data}
}
\description{
Generate a dygraph of journal paper percentages, by country and year
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
\dontrun{
pubmed_query_string <- paste(
  "passion [Title/Abstract]",
  "AND Dualistic Model of Passion [Text Word]"
)

save_process_pubmed_batch(
  pubmed_query_string,
  year_low = 2021,
  year_high = 2023
)
data <- read_bind_all_data()
dygraph_country_year(data)
}
\dontshow{
unlink("easyPubMed_data_01.txt")
unlink("articles_2021_2023.rds")
setwd(.old_wd)
}
}
