% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_address.R
\name{extract_split_address}
\alias{extract_split_address}
\title{Extracts correct component from a splitted affiliation address}
\usage{
extract_split_address(splitted.address, string)
}
\arguments{
\item{splitted.address}{The address splitted with \link[stringr:str_split]{stringr::str_split}.}

\item{string}{The keyword to determine which string component to keep.}
}
\description{
Extracts correct component from a splitted affiliation address
}
\examples{
address <- c(
  "Department of Psychology",
  " Cornell University",
  " Ithaca",
  " New York 14853-7601."
)

extract_split_address(address, "University")
extract_split_address(address, "Department")
}
