% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_affiliation.R
\name{get_affiliation}
\alias{get_affiliation}
\title{Get affiliations}
\usage{
get_affiliation(address, info = "university")
}
\arguments{
\item{address}{The address to parse.}

\item{info}{The information to extract, one of c("university", "department").}
}
\description{
Get affiliations
}
\examples{
address <- c(
  "Department of Psychology, Cornell University, Ithaca, New York 14853-7601.",
  "Dipartimento di Psicologia Generale, Università di Padova, Italy.",
  "Universität Mannheim, Federal Republic of Germany.",
  "Département de psychologie, Universit&#xe9; du Québec à Montréal, Canada."
)

get_affiliation(address, "department")
get_affiliation(address, "university")

}
