% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get_country}
\alias{get_country}
\title{Get country}
\usage{
get_country(address)
}
\arguments{
\item{address}{The address to parse.}
}
\description{
Get country
}
\examples{
\dontrun{
\dontshow{
.old_wd <- setwd("man")
}
d.fls <- batch_pubmed_download2(
  pubmed_query_string = paste(
    "passion [Title/Abstract]",
    "AND Dualistic Model of Passion [Text Word]",
    "AND ('2023/01/01' [Date - Publication] : '2023/12/31' [Date - Publication])"
  ),
  year_low = 2023,
  year_high = 2023
)
articles.df <- all_articles_to_df(d.fls)
articles.df2 <- add_affiliation(articles.df)
articles.df3 <- match_university(articles.df2)
get_country(articles.df3$address)
\dontshow{
setwd(.old_wd)
}
}
}
