% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal_field.R
\docType{data}
\name{journal_field}
\alias{journal_field}
\title{List of academic journals and corresponding fields}
\format{
A data frame with 25 rows and 3 variables:
\describe{
\item{journal}{academic journal}
\item{journal_short}{short name for the journal}
\item{field}{the field of research}
\item{original_journal}{Whether it is part of one of the six original journals}
...
}
}
\usage{
journal_field
}
\description{
List of academic journals and corresponding fields
}
\keyword{datasets}
