% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_university.R
\name{match_university}
\alias{match_university}
\title{Match list of universities to countries in a dataframe}
\usage{
match_university(data)
}
\arguments{
\item{data}{The dataframe to use for matching.}
}
\description{
Match list of universities to countries in a dataframe
}
