% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_vlookup.R
\name{partial_vlookup}
\alias{partial_vlookup}
\title{Like Excel vlookup/grep, both both ways}
\usage{
partial_vlookup(pattern, lookup_vector)
}
\arguments{
\item{pattern}{The pattern to compare.}

\item{lookup_vector}{The dictionnary in which to look for the pattern.}
}
\description{
Like Excel vlookup/grep, both both ways
}
\examples{
address <- c(
  "Department of Psychology, Cornell University, Ithaca, New York 14853-7601.",
  "Dipartimento di Psicologia Generale, Università di Padova, Italy.",
  "Universität Mannheim, Federal Republic of Germany.",
  "Département de psychologie, Université du Québec à Montréal, Canada."
)
partial_vlookup(address, universities$university)
uni <- c(
  "Cornell University", "Università di Padova",
  "Universität Mannheim", "Université du Québec à Montréal"
)
partial_vlookup(uni, universities$university)
}
