% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bind_all_data.R
\name{read_bind_all_data}
\alias{read_bind_all_data}
\title{Read local pubmedDashboard data files and bind them in a single dataframe}
\usage{
read_bind_all_data(data_folder = "data")
}
\arguments{
\item{data_folder}{The folder in which the data lives}
}
\description{
Read local pubmedDashboard data files and bind them in a single dataframe
}
