% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_dashboard.R
\name{render_dashboard}
\alias{render_dashboard}
\title{Render complete pubmedDashboard dashboard}
\usage{
render_dashboard(
  file_name = "dashboard",
  title = "title",
  author = "author",
  pubmed_query_string = "",
  journal = NULL,
  year_low = 2023,
  year_high = 2023,
  month_low = "01",
  month_high = 12,
  day_low = "01",
  day_high = 31,
  data_folder = "data",
  batch_size = 5000,
  api_key = NULL,
  verbose = TRUE,
  query_pubmed = FALSE,
  tab_continent = TRUE,
  tab_continent_year = TRUE,
  tab_continent_journal = TRUE,
  tab_country = TRUE,
  tab_country_journal = TRUE,
  tab_psychology = FALSE,
  tab_economics = FALSE,
  tab_general = FALSE,
  tab_figure1 = FALSE,
  tab_missing = TRUE
)
}
\arguments{
\item{file_name}{Desired file name.}

\item{title}{Desired dashboard title.}

\item{author}{Desired displayed dashboard author.}

\item{pubmed_query_string}{The PubMed query string.}

\item{journal}{The list of desired journals.}

\item{year_low}{The year the data should start.}

\item{year_high}{The year the data should end.}

\item{month_low}{The year the data should start.}

\item{month_high}{The year the data should end.}

\item{day_low}{The year the data should start.}

\item{day_high}{The year the data should end.}

\item{data_folder}{Where to save the data.}

\item{batch_size}{The download batch size.}

\item{api_key}{The api key for faster processing (optional).}

\item{verbose}{Whether to include progress messages.}

\item{query_pubmed}{Whether to query pubmed. This must be set to TRUE
explicitely to avoid long operations. When the data is already downloaded
and available in the data folder, this step is unnecessary.}

\item{tab_continent}{Whether to render the "Continent" tab.}

\item{tab_continent_year}{Whether to render the "Continent by year" tab.}

\item{tab_continent_journal}{Whether to render the "Continent by journal" tab.}

\item{tab_country}{Whether to render the "Country" tab.}

\item{tab_country_journal}{Whether to render the "Country by journal" tab.}

\item{tab_psychology}{Whether to render the "Psychology" tab.}

\item{tab_economics}{Whether to render the "Economics" tab.}

\item{tab_general}{Whether to render the "General" tab.}

\item{tab_figure1}{Whether to render the "Figure 1" tab.}

\item{tab_missing}{Whether to render the "Missing" tab.}
}
\description{
Render complete pubmedDashboard dashboard
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
\dontrun{
render_dashboard(
  file_name = "my_dashboard",
  title = "Wonderful Dashboard",
  author = "Rémi Thériault",
  pubmed_query_string = "passion [Title/Abstract]",
  journal = c("Journal of Personality and Social Psychology", "Health Psychology"),
  year_low = 2023,
  year_high = 2023,
  query_pubmed = TRUE,
  tab_figure1 = TRUE
)
}
\dontshow{
unlink("easyPubMed_data_01.txt")
setwd(.old_wd)
}
}
