% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_process_pubmed_batch.R
\name{save_process_pubmed_batch}
\alias{save_process_pubmed_batch}
\title{Mega function to process and save PubMed data}
\usage{
save_process_pubmed_batch(
  pubmed_query_string,
  journal = "",
  year_low = 2023,
  year_high = 2023,
  data_folder = "data",
  batch_size = 5000,
  api_key = NULL,
  verbose = TRUE
)
}
\arguments{
\item{pubmed_query_string}{The PubMed query string.}

\item{journal}{The list of desired journals.}

\item{year_low}{The year the data should start.}

\item{year_high}{The year the data should end.}

\item{data_folder}{Where to save the data.}

\item{batch_size}{The download batch size.}

\item{api_key}{The api key for faster processing (optional).}

\item{verbose}{Whether to include progress messages.}
}
\description{
Mega function to process and save PubMed data
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
\dontrun{
pubmed_query_string <- paste(
  "passion [Title/Abstract]",
  "AND Dualistic Model of Passion [Text Word]"
)

save_process_pubmed_batch(
  pubmed_query_string,
  year_low = 2023,
  year_high = 2023
)
}
\dontshow{
unlink("easyPubMed_data_01.txt")
setwd(.old_wd)
}
}
