% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_continent_year.R
\name{scatter_continent_year}
\alias{scatter_continent_year}
\title{Generate table of journal paper percentages, by continent and year}
\usage{
scatter_continent_year(
  data,
  method = "lm",
  plotly = TRUE,
  citation = NULL,
  citation_size = 15,
  ...
)
}
\arguments{
\item{data}{The processed dataframe of data}

\item{method}{Which method to use for the regression line, either "lm" (default) or "loess".}

\item{plotly}{Logical, whether to use plotly for dynamic data visualization.}

\item{citation}{Optionally, a citation to add as a footer.}

\item{citation_size}{Font size of the citation.}

\item{...}{Further arguments passed to \link[rempsyc:nice_scatter]{rempsyc::nice_scatter}}
}
\description{
Generate table of journal paper percentages, by continent and year
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
\dontrun{
pubmed_query_string <- paste(
  "passion [Title/Abstract]",
  "AND Dualistic Model of Passion [Text Word]"
)

save_process_pubmed_batch(
  pubmed_query_string,
  year_low = 2022,
  year_high = 2023
)
data <- read_bind_all_data()
suppressWarnings(scatter_continent_year(data))
}
\dontshow{
unlink("easyPubMed_data_01.txt")
setwd(.old_wd)
}
}
