% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_figure1.R
\name{scatter_figure1}
\alias{scatter_figure1}
\title{Generate table of journal paper percentages, by continent and year}
\usage{
scatter_figure1(data, method = "lm", original = TRUE, plotly = TRUE, ...)
}
\arguments{
\item{data}{The processed dataframe of data}

\item{method}{Which method to use for the regression line, either "lm" (default) or "loess".}

\item{original}{Logical; if \code{TRUE}, attempts to mimic Arnett's (2008) Figure 1 in style.}

\item{plotly}{Logical, whether to use plotly for dynamic data visualization.}

\item{...}{Further arguments passed to \link[rempsyc:nice_scatter]{rempsyc::nice_scatter}}
}
\description{
Generate table of journal paper percentages, by continent and year
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
\dontrun{
pubmed_query_string <- paste(
  "Developmental Psychology [Journal]",
  "OR Journal of Personality and Social Psychology [Journal]",
  "OR Journal of Abnormal Psychology [Journal]",
  "OR Journal of Family Psychology [Journal]",
  "OR Health Psychology [Journal]",
  "OR Journal of Educational Psychology [Journal]"
)

save_process_pubmed_batch(
  pubmed_query_string,
  year_low = 2023,
  year_high = 2023
)
data <- read_bind_all_data()
scatter_figure1(data)
}
\dontshow{
unlink("easyPubMed_data_01.txt")
setwd(.old_wd)
}
}
