% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_journal_continent_year.R
\name{scatter_journal_continent_year}
\alias{scatter_journal_continent_year}
\title{Generate table of journal paper percentages, by continent and year}
\usage{
scatter_journal_continent_year(data, method = "lm")
}
\arguments{
\item{data}{The processed dataframe of data}

\item{method}{Which method to use for the regression line, either "lm" (default) or "loess".}
}
\description{
Generate table of journal paper percentages, by continent and year
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
\dontrun{
pubmed_query_string <- paste(
  "passion [Title/Abstract]",
  "AND Dualistic Model of Passion [Text Word]"
)

save_process_pubmed_batch(
  pubmed_query_string,
  year_low = 2022,
  year_high = 2023
)
data <- read_bind_all_data()
suppressWarnings(scatter_journal_continent_year(data))
}
\dontshow{
unlink("easyPubMed_data_01.txt")
setwd(.old_wd)
}
}
