% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_address.R
\name{split_address}
\alias{split_address}
\title{Split affiliation address into separate strings}
\usage{
split_address(address, string)
}
\arguments{
\item{address}{The address to split.}

\item{string}{The character used to split the address.}
}
\description{
Split affiliation address into separate strings
}
