% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_journals_continents.R
\name{table_journal_count}
\alias{table_journal_count}
\title{Count number of papers per journal, with year range}
\usage{
table_journal_count(data, datatable = TRUE)
}
\arguments{
\item{data}{The processed dataframe of data}

\item{datatable}{Whether to output a \link[DT:datatable]{DT::datatable} HTML table widget
instead of a regular dataframe (defaults to TRUE).}
}
\description{
Count number of papers per journal, with year range
}
