% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_missing_country.R
\name{table_missing_country}
\alias{table_missing_country}
\title{Generate table of journal paper percentages, by country}
\usage{
table_missing_country(data, datatable = TRUE)
}
\arguments{
\item{data}{The processed dataframe of data}

\item{datatable}{Whether to output a \link[DT:datatable]{DT::datatable} HTML table widget
instead of a regular dataframe (defaults to TRUE).}
}
\description{
Generate table of journal paper percentages, by country
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
\dontrun{
pubmed_query_string <- paste(
  "passion [Title/Abstract]",
  "AND Dualistic Model of Passion [Text Word]"
)

save_process_pubmed_batch(
  pubmed_query_string,
  year_low = 2023,
  year_high = 2023
)
data <- read_bind_all_data()
table_missing_country(data)
}
\dontshow{
unlink("easyPubMed_data_01.txt")
setwd(.old_wd)
}
}
