% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universities.R
\docType{data}
\name{universities}
\alias{universities}
\title{A data frame of university and corresponding country}
\format{
A data frame with 9420 rows and 2 variables:
\describe{
\item{country_code}{the country code}
\item{university}{the university}
...
}
}
\source{
\url{https://raw.githubusercontent.com/endSly/world-universities-csv/master/world-universities.csv}
}
\usage{
universities
}
\description{
Obtained from GitHub, and then modified with minor improvements and more
universities.
}
\keyword{datasets}
