% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_states.R
\docType{data}
\name{us_states}
\alias{us_states}
\title{List of US states taken from the package \code{countrycode}}
\format{
A data frame with 50 rows and 3 variables:
\describe{
\item{state.name}{the name of the state}
\item{state.abb}{the name of the abbreviation}
\item{state.regex}{the regex for that state}
...
}
}
\usage{
us_states
}
\description{
List of US states taken from the package \code{countrycode}
}
\keyword{datasets}
