% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle_continent.R
\name{waffle_continent}
\alias{waffle_continent}
\title{Generate a waffle chart of journal paper percentages, by continent (each square = 1\% of data)}
\usage{
waffle_continent(data, citation = NULL)
}
\arguments{
\item{data}{The processed dataframe of data}

\item{citation}{Optionally, a citation to add as a footer.}
}
\description{
Generate a waffle chart of journal paper percentages, by continent (each square = 1\% of data)
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
\dontrun{
pubmed_query_string <- paste(
  "passion [Title/Abstract]",
  "AND Dualistic Model of Passion [Text Word]"
)

save_process_pubmed_batch(
  pubmed_query_string,
  year_low = 2023,
  year_high = 2023
)
data <- read_bind_all_data()
waffle_continent(data)
}
\dontshow{
unlink("easyPubMed_data_01.txt")
setwd(.old_wd)
}
}
