% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle_continent_journal.R
\name{waffle_continent_journal}
\alias{waffle_continent_journal}
\title{Generate a waffle chart of journal paper percentages, by continent (each square = 1\% of data)}
\usage{
waffle_continent_journal(data, citation, journal_abbreviation = TRUE)
}
\arguments{
\item{data}{The processed dataframe of data}

\item{citation}{Optionally, a citation to add as a footer.}

\item{journal_abbreviation}{Logical, whether to use the journal abbreviation
to fit the entire plot, otherwise some journal names can be quite long and
accordingly be cropped.}
}
\description{
Generate a waffle chart of journal paper percentages, by continent (each square = 1\% of data)
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
\dontrun{
pubmed_query_string <- paste(
  "passion [Title/Abstract]",
  "AND Dualistic Model of Passion [Text Word]"
)

save_process_pubmed_batch(
  pubmed_query_string,
  year_low = 2023,
  year_high = 2023
)
data <- read_bind_all_data()
waffle_continent_journal(data)
}
\dontshow{
unlink("easyPubMed_data_01.txt")
setwd(.old_wd)
}
}
