% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle_country.R
\name{waffle_country}
\alias{waffle_country}
\title{Generate a waffle plot made of country flags}
\usage{
waffle_country(data, citation)
}
\arguments{
\item{data}{The processed dataframe of data}

\item{citation}{Optionally, a citation to add as a footer.}
}
\description{
Generate a waffle plot made of country flags
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
\dontrun{
pubmed_query_string <- paste(
  "passion [Title/Abstract]",
  "AND Dualistic Model of Passion [Text Word]"
)

save_process_pubmed_batch(
  pubmed_query_string,
  year_low = 2023,
  year_high = 2023
)
data <- read_bind_all_data()
waffle_country(data)
}
\dontshow{
unlink("easyPubMed_data_01.txt")
unlink("articles_2023_2023.rds")
setwd(.old_wd)
}
}
