% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle_flags.R
\name{waffle_flags}
\alias{waffle_flags}
\title{Generate a waffle plot made of country flags}
\usage{
waffle_flags(in_map_var, len_x = NA, na_flag = "ac")
}
\arguments{
\item{in_map_var}{The data}

\item{len_x}{Length on the x-axis}

\item{na_flag}{Which flag to use for missing values}
}
\description{
Generate a waffle plot made of country flags
}
