% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle_journal_continent.R
\name{waffle_journal_continent}
\alias{waffle_journal_continent}
\title{Generate a waffle chart of journal paper percentages, by continent (each square = 1\% of data)}
\usage{
waffle_journal_continent(data)
}
\arguments{
\item{data}{The processed dataframe of data}
}
\description{
Generate a waffle chart of journal paper percentages, by continent (each square = 1\% of data)
}
\examples{
\dontrun{
\dontshow{
.old_wd <- setwd("man")
}
pubmed_query_string <- paste(
  "passion [Title/Abstract]",
  "AND Dualistic Model of Passion [Text Word]"
)

save_process_pubmed_batch(
  pubmed_query_string,
  year_low = 2023,
  year_high = 2023
)
data <- read_bind_all_data()
waffle_journal_continent(data)
\dontshow{
setwd(.old_wd)
}
}
}
