% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_capitals.R
\docType{data}
\name{world_capitals}
\alias{world_capitals}
\title{List of world capitals taken from the package \code{maps}}
\format{
A data frame with 259 rows and 6 variables:
\describe{
\item{name}{the name of the capital}
\item{country.etc}{the country of the capital}
\item{pop}{population of the capital}
\item{lat}{latitude of the capital}
\item{long}{longitude of the capital}
\item{capital}{whether it is a capital}
...
}
}
\usage{
world_capitals
}
\description{
List of world capitals taken from the package \code{maps}
}
\keyword{datasets}
