% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_affiliations.R
\name{add_affiliation}
\alias{add_affiliation}
\title{Add affiliations to a dataframe}
\usage{
add_affiliation(data)
}
\arguments{
\item{data}{The dataframe on which to add affiliations (department and
university).}
}
\description{
Add affiliations to a dataframe
}
