% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_articles_to_df.R
\name{all_articles_to_df}
\alias{all_articles_to_df}
\title{Convert list of PubMed XLM files to dataframe}
\usage{
all_articles_to_df(d.fls)
}
\arguments{
\item{d.fls}{The list of XLM PubMed data.}
}
\description{
Convert list of PubMed XLM files to dataframe
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
\dontrun{
pubmed_query_string <- paste(
  "passion [Title/Abstract]",
  "AND Dualistic Model of Passion [Text Word]",
  "AND ('2023/01/01' [Date - Publication] : '2023/12/31' [Date - Publication])"
)
d.fls <- batch_pubmed_download2(
  pubmed_query_string = pubmed_query_string,
  year_low = 2023,
  year_high = 2023,
  data_folder = ""
)
articles.df <- all_articles_to_df(d.fls)
articles.df[5, ]
}
\dontshow{
unlink("easyPubMed_data_01.txt")
setwd(.old_wd)
}
}
