% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_university.R
\name{match_university}
\alias{match_university}
\title{Match list of universities to countries in a dataframe}
\usage{
match_university(data)
}
\arguments{
\item{data}{The dataframe to use for matching.}
}
\description{
Match list of universities to countries in a dataframe
}
\examples{
\dontrun{
d.fls <- batch_pubmed_download2(
  pubmed_query_string = paste(
    "passion [Title/Abstract]",
    "AND Dualistic Model of Passion [Text Word]",
    "AND ('2023/01/01' [Date - Publication] : '2023/12/31' [Date - Publication])"
  )
)
articles.df <- all_articles_to_df(d.fls)
articles.df2 <- add_affiliation(articles.df)
articles.df3 <- match_university(articles.df2)
articles.df3[5, ]
}
}
