% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_vlookup.R
\name{partial_vlookup}
\alias{partial_vlookup}
\title{Like Excel vlookup/grep, both both ways}
\usage{
partial_vlookup(pattern, lookup_vector)
}
\arguments{
\item{pattern}{The pattern to compare.}

\item{lookup_vector}{The dictionnary in which to look for the pattern.}
}
\description{
Like Excel vlookup/grep, both both ways
}
